requirements.txt
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.29.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
optuna>=3.0.0
pyyaml>=6.0
streamlit>=1.20.0
tensorboard>=2.10.0
imageio>=2.30.0
tabulate>=0.9.0